+++ maser/services/helio/hfc/__init__.py
@@ -1,3 +1,0 @@
-
-from maser.services.helio.hfc.subparser import add_hfcviewer_subparser
-from maser.services.helio.hfc.hfcviewer import main as hfcviewer